rewrote the generation of shellpairs to be lockfree
